Reorganised contants a little, attempted to add context click to map window. Didnt work. Added commands to send waypoints to phone, send strings from phone to server. Can start/stop waypoint driver from map screen. Waypoint driver calculates the alteration to make to the current heading to get to the waypoints. Next is making it actually control the ioio with override method

+++ trunk/androthumper/Android_CAR/src/android/ioio/car/listeners/MyCompassListener.java
@@ -2,5 +2,5 @@ package android.ioio.car.listeners;
 
 public interface MyCompassListener {
 
-	public void gotCompassHeading(int heading);
+	public void gotCompassHeading(float heading);
 }
